Moving common module codec to common package to resolve import cycle issue.
